Move `Target` to the top level
