+++ internal/services/input_button.go
@@ -3,6 +3,7 @@ package services
 import (
 	"context"
 
+	ga "saml.dev/gome-assistant"
 	"saml.dev/gome-assistant/websocket"
 )
 
@@ -24,13 +25,13 @@ func (ib InputButton) Press(entityID string) (websocket.Message, error) {
 	ctx := context.TODO()
 	return ib.service.CallService(
 		ctx, "input_button", "press",
-		nil, EntityTarget(entityID),
+		nil, ga.EntityTarget(entityID),
 	)
 }
 
 func (ib InputButton) Reload() (websocket.Message, error) {
 	ctx := context.TODO()
 	return ib.service.CallService(
-		ctx, "input_button", "reload", nil, Target{},
+		ctx, "input_button", "reload", nil, ga.Target{},
 	)
 }
