+++ internal/services/services.go
@@ -3,26 +3,9 @@ package services
 import (
 	"context"
 
+	ga "saml.dev/gome-assistant"
 	"saml.dev/gome-assistant/websocket"
 )
-
-// Target represents the target of the service call, if applicable.
-type Target struct {
-	EntityID string `json:"entity_id,omitempty"`
-	DeviceID string `json:"device_id,omitempty"`
-}
-
-func EntityTarget(entityID string) Target {
-	return Target{
-		EntityID: entityID,
-	}
-}
-
-func DeviceTarget(deviceID string) Target {
-	return Target{
-		DeviceID: deviceID,
-	}
-}
 
 type Service interface {
 	Call(
@@ -30,6 +13,6 @@ type Service interface {
 	) (websocket.Message, error)
 
 	CallService(
-		ctx context.Context, domain string, service string, serviceData any, target Target,
+		ctx context.Context, domain string, service string, serviceData any, target ga.Target,
 	) (websocket.Message, error)
 }
